Refatorado tabelas de classes de documento
