+++ src/main/kotlin/document/persistency/tables/civilRegistryDocumentTable.kt
@@ -1,0 +1,10 @@
+package document.persistency.tables
+
+import org.jetbrains.exposed.sql.Table
+
+object civilRegistryDocumentTable : Table("civil_registry_document") {
+    val id = char("id", 44).references(documentTable.id)
+    val registration = char("registration", 32)
+
+    override val primaryKey = PrimaryKey(id, name = "pk_civil_registry_document_id")
+}
