+++ src/main/kotlin/document/persistency/tables/grandparentTable.kt
@@ -1,0 +1,13 @@
+package document.persistency.tables
+
+import org.jetbrains.exposed.sql.Table
+
+object grandparentTable : Table("grandparent") {
+    val id = char("id", 32)
+    val personId = char("person_id",32)
+    val name = varchar("name", 120)
+    val UF = enumeration("uf", document.UF::class).nullable()
+    val Municipality = varchar("municipality", 80).nullable()
+
+    override val primaryKey = PrimaryKey(id, name = "pk_grandparent_id")
+}
