+++ src/main/kotlin/document/persistency/tables/observationsAndRegisteringTable.kt
@@ -1,0 +1,10 @@
+package document.persistency.tables
+
+import org.jetbrains.exposed.sql.Table
+
+object observationsAndRegisteringTable : Table("observations_and_registering") {
+    val id = char("id", 44).references(civilRegistryDocumentTable.id)
+    val text = text("text")
+
+    override val primaryKey = PrimaryKey(id, name = "pk_observations_and_registering_id")
+}
